Adds automated deployment versioning
Introduces automated versioning to the deployment process.

It reads the version from `package.json`, tags the built image with this version, and avoids unnecessary rebuilds if the current version is already deployed to the cluster.

A `--force` flag is added to bypass the version check and force a rebuild.

The `dates.sh` script retrieves the latest modified date of any file.

Fixes a defect in Keycloak Admin Service that was returning null.